More cleaning + a fix to PUT request 201 (include Content-Length=0)

+++ includes/ServerConfig.hpp
@@ -41,7 +41,7 @@ class ServerConfig {
   std::vector<Listen> &getListens();
   std::vector<std::string> &getServerNames();
   std::vector<ServerConfig> &getLocations();
-  int &getClientMaxBodySize();
+  size_t &getClientMaxBodySize();
   std::string &getRoot();
   std::map<int, std::string> &getErrorCodes();
   std::vector<std::string> &getIndexes();
@@ -56,7 +56,7 @@ class ServerConfig {
   std::vector<std::string> server_name_;
   std::vector<ServerConfig> locations_;
   std::string uri_;
-  int client_max_body_size_;
+  size_t client_max_body_size_;
   std::string root_;
   std::vector<std::string> methods_;
   std::map<int, std::string> error_codes_;
